// frontend/src/api/models/handlerAddStockTakingItemRequest.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Request body for adding a product to stock taking
 */
export interface HandlerAddStockTakingItemRequest {
  product_code: string
  product_id: string
  product_name: string
  system_quantity: number
  unit: string
  unit_cost: number
}